requirements.txt
Flask
google-cloud-speech
pyttsx3
google-generativeai
python-dotenv